+++ apps/mobile/App.tsx
@@ -264,6 +264,7 @@ import { MedicalSupplyRequestScreen } from "./screens/MedicalSupplyRequestScreen
 import { MedicalSuppliesSetupRoute } from "./screens/MedicalSuppliesSetupScreen";
 import { OutAboutScreen } from "./screens/OutAboutScreen";
 import { GoingOutScreen } from "./screens/GoingOutScreen";
+import { MyCardScreen } from "./screens/MyCardScreen";
 import { MyToolsScreen } from "./screens/MyToolsScreen";
 import { MY_TOOLS_HINT } from "./lib/myToolsShared";
 import {
@@ -4856,6 +4857,7 @@ function AppTabs({
       Support: "Support",
       OutAbout: "Out & About",
       GoingOut: "Going Out",
+      MyCard: "My Card",
       MyTools: "My tools",
       Reminders: "Reminders",
       Hydration: "My Hydration",
@@ -5116,6 +5118,7 @@ function AppTabs({
             <AppStack.Screen name="IbdAtWork">{() => <IbdAtWorkScreen />}</AppStack.Screen>
             <AppStack.Screen name="OutAbout">{() => <OutAboutScreen />}</AppStack.Screen>
             <AppStack.Screen name="GoingOut">{() => <GoingOutScreen userId={user.id} />}</AppStack.Screen>
+            <AppStack.Screen name="MyCard">{() => <MyCardScreen userId={user.id} />}</AppStack.Screen>
             <AppStack.Screen name="MyTools">{() => <MyToolsScreen />}</AppStack.Screen>
             <AppStack.Screen name="LatestNews">{() => <LatestNewsScreen user={user} />}</AppStack.Screen>
             <AppStack.Screen name="Account">